requirements.txt
asgiref==3.4.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-modeltranslation==0.16.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
